requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asn1crypto==1.5.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cryptography==40.0.1
cssselect2==0.7.0
cycler==0.12.1
dateutils==0.6.12
deprecation==2.1.0
deta==1.2.0
distlib==0.3.6
dj-database-url==1.3.0
Django==4.1.7
django-filter==23.1
django-heroku==0.3.1
django-stubs==1.16.0
django-stubs-ext==0.8.0
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==24.3.0
favicon==0.7.0
filelock==3.10.7
firebase-admin==6.5.0
fonttools==4.49.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.4.1
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==20.1.0
h11==0.14.0
heroku==0.1.4
htbuilder==0.6.2
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.4
imbalanced-learn==0.12.0
imblearn==0.0
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==4.9.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
mypy==1.2.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==1.26.0
oscrypto==1.3.0
packaging==23.2
pandas==2.1.1
patsy==0.5.6
Pillow==9.4.0
pisa==3.0.33
platformdirs==3.2.0
plotly==5.20.0
plotly-express==0.4.1
postgrest==0.16.2
posts==0.2.2
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.5
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.8.0
pymdown-extensions==10.7.1
pyodbc==5.1.0
pyparsing==3.1.2
pypdf==3.7.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
realtime==1.0.2
referencing==0.33.0
reportlab==3.6.12
requests==2.28.2
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.3
st-annotated-text==4.0.1
st-login-form==0.2.2
st-pages==0.4.5
statsmodels==0.14.1
storage3==0.7.3
streamlit==1.32.2
streamlit-authenticator==0.3.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-cookies-manager==0.2.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-login-auth-ui==0.2.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.12
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
supabase==2.4.0
supafunc==0.3.3
svglib==1.5.1
tenacity==8.2.3
threadpoolctl==3.3.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
trycourier==5.0.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
typing_extensions==4.10.0
tzdata==2023.2
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
utils==1.0.2
validators==0.22.0
virtualenv==20.21.0
watchdog==4.0.0
webencodings==0.5.1
websockets==11.0.3
whitenoise==6.4.0
xhtml2pdf==0.2.9
zipp==3.17.0